refactor(desktop): extract chat constants to chat-constants.ts
move pure const lookup tables and small helpers (PROVIDER_LABEL,
PROVIDER_TEXT, EFFORT_*, MODEL_COST, FAMILY_LABEL, TIER_*, VERBOSITY_*,
modelLabel, modelFamily, modelTier, modelWeight) from ChatInput.tsx
into a dedicated chat-constants.ts module. ChatInput drops from 844
to 739 lines. no behavior change.

note: AuthRequiredCallout has a separate, intentionally-different
PROVIDER_LABEL (lowercase values) — left untouched; unifying display
casing is out of scope for this refactor.

Closes #396
